Everything is np.int32 now.

+++ src/dataset.py
@@ -155,11 +155,11 @@ class VectorGetter():
 
 
         y = self.y_label_encoder.transform(y).reshape(-1, 1)
-        y = self.y_onehot_encoder.transform(y).todense()
+        y = self.y_onehot_encoder.transform(y).todense(dtype=np.int32)
 
 
 
-        X = np.array(X)
+        X = np.array(X, dtype=np.int32)
         # print(len(y), "individual tracks loaded!")
 
         return X, y
@@ -193,8 +193,8 @@ class VectorGetter():
 
 
         y = self.y_label_encoder.transform(y).reshape(-1, 1)
-        y = self.y_onehot_encoder.transform(y).todense()
-        X = np.array(X)
+        y = self.y_onehot_encoder.transform(y).todense(dtype=np.int32)
+        X = np.array(X, dtype=np.int32)
 
         X_train, X_test, y_train, y_test = train_test_split(X, y)
         return X_train, X_test, y_train, y_test
